requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
# psycopg2==2.9.9
psycopg2-binary==2.9.9
python-jose
passlib[bcrypt]
python-dotenv
pydantic[email]
email-validator
python-multipart
bcrypt
fastapi-mail
request
httpx